Implemented logic to solve puzzle
The logic will attempt to solve the puzzle in 9 different methods. Each
method to solve the puzzle tries to inject a "zero" cell starting at a
different number each time. Currently, the puzzle being used is too
difficult to solve. Perhaps more logic is needed, or a simpler puzzle
should be used for the time being.

I'll add more comments to the code later when I have time!