Remaining logic to solve puzzles
The app now successfully solves puzzles with up to 42 zeros in a
reasonable amount of time (anywhere between 1 - 10 seconds) using a
combination of 5 different techniques. The sudoku_values.txt must be
placed in the root level of the Sudoku project in your workspace in
order for the program to find it. I had to keep changing the values, so
I implemented the BufferedReader to speed things up. I can most likely
speed this app up further in order to solve puzzles with >50 zeros, but
I'll have to make it a side project to incorporate multiple threads.