Re-implemented processor logic
The processor didn't work how I intended it to, so I reimplemnted it
using the original recursion idea. It seems reliable and can complelely
solve Sudoku puzzles in a short amount of time.

Don't forget to add the sudoku_values.txt to the Sudoku project's folder
in Eclipse, otherwise it won't be recognized.